+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - ReleaseDate
 ### Added
+
+- Added the ability to match non-`Send` arguments with `withf_st`
+  ([#93](https://github.com/asomers/mockall/pull/93))
+
 ### Changed
 ### Fixed
 ### Removed